Split time-dependent field classes by responsibility
