+++ source/source_estate/CMakeLists.txt
@@ -24,6 +24,8 @@ list(APPEND objects
     module_pot/pot_xc_fdm.cpp
     module_pot/pot_cosikr.cpp
     module_pot/td_field.cpp
+    module_pot/td_field_profiles.cpp
+    module_pot/td_field_manager.cpp
     module_charge/chgmixing.cpp
     module_charge/gint_precision_controller.cpp
     module_charge/charge.cpp
