+++ source/source_estate/module_pot/H_TDDFT_pw.cpp
@@ -5,7 +5,7 @@
 #include "source_base/tool_quit.h"
 #include "source_io/module_efield/td_efield_io.h"
 #include "source_io/module_parameter/parameter.h"
-#include "td_field.h"
+#include "td_field_manager.h"
 
 namespace elecstate
 {
