+++ source/source_estate/test/CMakeLists.txt
@@ -74,7 +74,8 @@ AddTest(
 AddTest(
   TARGET MODULE_ESTATE_potentials_new
   LIBS parameter  base device planewave_serial
-  SOURCES potential_new_test.cpp td_field_test.cpp ../module_pot/potential_new.cpp ../module_pot/td_field.cpp
+  SOURCES potential_new_test.cpp td_field_test.cpp ../module_pot/potential_new.cpp
+    ../module_pot/td_field.cpp ../module_pot/td_field_profiles.cpp ../module_pot/td_field_manager.cpp
 )
 
 AddTest(
