+++ source/source_estate/test/td_field_test.cpp
@@ -1,6 +1,6 @@
-#include "source_estate/module_pot/td_field.h"
-
 #include "source_base/constants.h"
+#include "source_estate/module_pot/td_field_manager.h"
+#include "source_estate/module_pot/td_field_profiles.h"
 #include "source_io/module_parameter/input_parameter.h"
 
 #include "gtest/gtest.h"
