+++ source/source_io/test/CMakeLists.txt
@@ -37,7 +37,9 @@ AddTest(
 AddTest(
   TARGET MODULE_IO_td_efield_io_test
   LIBS parameter base device
-  SOURCES td_efield_io_test.cpp ../module_efield/td_efield_io.cpp ../../source_estate/module_pot/td_field.cpp
+  SOURCES td_efield_io_test.cpp ../module_efield/td_efield_io.cpp
+    ../../source_estate/module_pot/td_field.cpp ../../source_estate/module_pot/td_field_profiles.cpp
+    ../../source_estate/module_pot/td_field_manager.cpp
 )
 
 AddTest(
